feat: :sparkles: changed user activities
added message activities and changed the design of the activities
